Il faut calmer gcc pour s'adapter à une architecture LP64...
Dans les conversions (void*)<->(int) (linéarisation des pointeurs
dans les sauvegardes/restaurations du graphe dans/vers un fichier
binaire sur disque), il faut passer par un type "intptr_t" (défini
dans sys/types.h sur les compiltateurs C99) pour être sûr que le
cast se réalise avec tous les bits significatifs.

+++ anaconf/lib/binwrite.c
@@ -20,7 +20,7 @@ void bin_write (FILE *fpout, MOBJ *graph [])
     {
 	hdr.mobjhdr [i].objsiz = mobj_size (graph [i]) ;
 	hdr.mobjhdr [i].objcnt = mobj_count (graph [i]) ;
-	hdr.mobjhdr [i].listhead = (int) mobj_head (graph [i]) ;
+	hdr.mobjhdr [i].listhead = (intptr_t) mobj_head (graph [i]) ;
     }
 
     if (fwrite (&hdr, sizeof hdr, 1, fpout) != 1)
